Changing u_int64_t into size_t for using IHspContainer::retrieve method => compilation issue on MacOs where size_t != u_int64_t

+++ src/algo/hits/hsp/AlignmentGeneratorCmd.cpp
@@ -115,7 +115,7 @@ AlignmentGeneratorCmd::~AlignmentGeneratorCmd ()
 *********************************************************************/
 void AlignmentGeneratorCmd::execute ()
 {
-    u_int64_t nbHsp=0;
+    size_t nbHsp=0;
 
     IHspContainer::HSP* hsp = 0;
